requirements.txt
asgiref==3.7.2
backup==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.8
django-allauth==0.59.0
django-debug-toolbar==4.2.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.13.1
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient @ file:///C:/Users/zzxx1/Downloads/mysqlclient-1.4.6-cp39-cp39-win32.whl#sha256=7e0fb8aa09b57b3d92d8484f3ad507d1c7cc74de5b41946ea704d4d1907b0fe4
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
platformdirs==4.1.0
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.1.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
semantic-version==2.10.0
setuptools-rust==1.8.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.25.0
